Move parser to individual files
