+++ src/bin/main.c
@@ -9,7 +9,7 @@
 #include <unistd.h>
 #include <getopt.h>
 #include "../core/identifier.h"
-#include "../core/parser.h"
+#include "../core/parser/parser.h"
 #include "../core/compiler_xstate.h"
 #include "../core/error.h"
 #include "../core/str_builder.h"
