+++ src/core/compiler_xstate.h
@@ -1,5 +1,4 @@
-#ifndef LUCY_COMPILER_XSTATE_H_
-#define LUCY_COMPILER_XSTATE_H_
+#pragma once
 
 typedef struct CompileResult {
   bool success;
@@ -12,5 +11,3 @@ void xs_init(CompileResult*, int);
 void compile_xstate(CompileResult*, char*, char*);
 char* xs_get_js(CompileResult*);
 void destroy_xstate_result(CompileResult*);
-
-#endif
