+++ src/core/parser/invoke.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "../state.h"
+
+int parser_consume_invoke(State*);
