+++ src/core/parser/machine.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "../state.h"
+
+int parser_consume_machine(State*);
+int parser_consume_implicit_machine(State*, int);
