+++ src/core/parser/spawn.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "../node.h"
+#include "../state.h"
+
+int parser_consume_inline_spawn(State*, AssignExpression*);
