+++ src/core/parser/use.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "../state.h"
+
+int parser_consume_use(State*);
